Move UserContext into a new folder (src/components/contexts)
