+++ client/src/components/NavBar.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { UserContext } from './UserContext';
+import { UserContext } from './contexts/UserContext';
 import { Link as RouterLink } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 import {
