+++ client/src/pages/Landing/components/LoginForm.js
@@ -10,7 +10,7 @@ import {
 } from '@material-ui/core';
 import { Alert } from '@material-ui/lab';
 import CheckIcon from '@material-ui/icons/Check';
-import { UserContext } from '../../../components/UserContext';
+import { UserContext } from '../../../components/contexts/UserContext';
 
 let passwordMinLength = 6;
